tests/submission/PhilipTaylor/canvas/2d.type.extend.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.type.extend</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.type.html">type</a>.extend</h1>
<p class="desc">Interface methods can be added</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.context.2d.type">context.2d.type</a>
</li></ul>
</div>
<p class="notes">Defined in "Web IDL" (draft)
</p><p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="green-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("Interface methods can be added");
_addTest(function(canvas, ctx) {

window.CanvasRenderingContext2D.prototype.fillRectGreen = function (x, y, w, h)
{
    this.fillStyle = '#0f0';
    this.fillRect(x, y, w, h);
};
ctx.fillStyle = '#f00';
ctx.fillRectGreen(0, 0, 100, 50);
_assertPixel(canvas, 50,25, 0,255,0,255, "50,25", "0,255,0,255");


});
</script>

<div id="log"></div>
</body></html>
